Add validation for input field, refactor styles
