+++ ffm-java-22-1-front-end/src/Header.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import "./Css/Header.css"
+import "./css/Header.css"
 
 function Header() {
     return (
