+++ ffm-java-22-1-front-end/src/TodoGallery.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
-import {TodoModel} from "./Model/TodoModel";
+import {TodoModel} from "./model/TodoModel";
 import ToDoCard from "./ToDoCard";
-import "./Css/ToDoGallery.css"
+import "./css/ToDoGallery.css"
 
 type GalleryProps = {
     todoList: TodoModel[],
@@ -16,19 +16,30 @@ function TodoGallery(props: GalleryProps) {
     return (
         <div className={"todo-gallery"}>
             <div className={"todo-gallery-column"}>
-                <h2>Open</h2>
+                <div className={"todo-gallery-column-title"}>
+                    <h2>Open</h2>
+                    <p>(Items: {tasksOpen.length})</p>
+                </div>
+
+
             {tasksOpen.map((current) =>
                 <ToDoCard singleToDo={current} key={current.id} refreshPage={props.refreshPage}/>
             )}
             </div>
             <div className={"todo-gallery-column"}>
-                <h2>In Progress</h2>
+                <div className={"todo-gallery-column-title"}>
+                    <h2>In progress</h2>
+                    <p>(Items: {tasksInProgress.length})</p>
+                </div>
                 {tasksInProgress.map((current) =>
                     <ToDoCard singleToDo={current} key={current.id} refreshPage={props.refreshPage}/>
                 )}
             </div>
             <div className={"todo-gallery-column"}>
-                <h2>Done</h2>
+                <div className={"todo-gallery-column-title"}>
+                    <h2>Done</h2>
+                    <p>(Items: {tasksDone.length})</p>
+                </div>
                 {tasksDone.map((current) =>
                     <ToDoCard singleToDo={current} key={current.id} refreshPage={props.refreshPage}/>
                 )}
