+++ ffm-java-22-1-front-end/src/css/Header.css
@@ -1,7 +1,7 @@
 .header {
     font-family: Georgia;
     text-align: left;
-    padding: 60px;
+    padding: 1em;
     background: #1abc9c;
     color: white;
     font-size: 30px;
