+++ ffm-java-22-1-front-end/src/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import './Css/index.css';
+import './css/index.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
